Change modal to dialog.
